requirements.txt
streamlit
openai
langchain
faiss-cpu
tiktoken
python-dotenv
PyPDF2
python-docx